requirements.txt
os
argparse
torch
torchvision
time
pdb
numpy
tqdm==4.64.0
opencv-python==4.6.0.66
h5py==3.6.0
wandb==0.12.18
matplotlib==3.8.2
prettytable
efficientnet-pytorch==0.7.1
einops==0.7.0
warmup-scheduler
diffusers==0.11.1
lmdb 
vit-pytorch==1.6.5
positional-encodings
tensorflow
tensorflow_datasets
PyYAML==6.0.1
pickle
pillow==10.2.0
datetime
typing==3.7.4.3
typing_extensions==4.9.0
itertools

